code/api-monitor/api/requirements.txt
fastapi
uvicorn
requests
pyyaml
jsonschema
prometheus_client
apscheduler